Addressed some review comments
