+++ conda_forge_tick/auto_tick.xsh
@@ -26,7 +26,6 @@ $MIGRATORS = [
    # Noarch(pr_limit=10),
    # Pinning(pr_limit=1, removals={'perl'}),
    # Compiler(pr_limit=7),
-   ArchRebuild(pr_limit=7)
 ]
 
 def run(attrs, migrator, feedstock=None, protocol='ssh',
